+++ Sales_management/src/main/java/com/example/demo/entity/Status.java
@@ -28,7 +28,7 @@ public class Status implements Serializable {
 	@Id
 	@Column(name = "customer_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int customer_id;
+	private Long customer_id;
 
 	/**
 	 * ステータスID
